+++ Edutech-main/course-service/src/main/java/com/edutech/course_service/service/ModuloService.java
@@ -9,6 +9,7 @@ import com.edutech.course_service.model.Course;
 import com.edutech.course_service.model.Modulo;
 import com.edutech.course_service.repository.CourseRepository;
 import com.edutech.course_service.repository.ModuloRepository;
+import com.edutech.course_service.webclient.AuthClient;
 
 import jakarta.transaction.Transactional;
 
@@ -20,6 +21,9 @@ public class ModuloService {
 
     @Autowired
     private CourseRepository courseRepository;
+
+    @Autowired
+    private AuthClient authClient;
     
     // Obtener todos los módulos
     public List<Modulo> obtenerModulos() {
@@ -38,13 +42,18 @@ public class ModuloService {
     }
 
     // Crear nuevo módulo
-    public Modulo crearModulo(String titulo, String descripcion, Integer orden, Long cursoId) {
-        if (titulo == null || descripcion == null || orden == null || cursoId == null) {
+    public Modulo crearModulo(String authHeader, String titulo, String descripcion, Integer orden, Long cursoId) {
+        if (titulo == null || descripcion == null || orden == null || orden < 1 || cursoId == null) {
             throw new RuntimeException("Todos los campos son obligatorios");
         }
-        
+
         Course curso = courseRepository.findById(cursoId)
-        .orElseThrow(()-> new RuntimeException("Curso no encontrado Id: "+ cursoId));
+            .orElseThrow(() -> new RuntimeException("Curso no encontrado Id: " + cursoId));
+
+        if (!authClient.usuarioPuedeModificarCurso(authHeader, curso.getInstructorId())) {
+            throw new RuntimeException("No tienes permiso para agregar módulos a este curso");
+        }
+
         Modulo modulo = new Modulo();
         modulo.setTitulo(titulo);
         modulo.setDescripcion(descripcion);
@@ -54,12 +63,20 @@ public class ModuloService {
         return moduloRepository.save(modulo);
     }
 
-    // Actualizar módulo existente
-    public Modulo actualizarModulo(Long id, String titulo, String descripcion, Integer orden) {
-        Modulo modulo = moduloRepository.findById(id)
-        .orElseThrow(() -> new RuntimeException("Módulo no encontrado Id: " + id));
 
-        if (titulo != null && !titulo.trim().isEmpty()) {
+    // Actualizar módulo existente
+    public Modulo actualizarModulo(String authHeader, Long id, String titulo, String descripcion, Integer orden, Long cursoId) {
+        Modulo modulo = moduloRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Módulo no encontrado Id: " + id));
+
+        Course curso = courseRepository.findById(cursoId)
+            .orElseThrow(() -> new RuntimeException("Curso no encontrado"));
+
+        if (!authClient.usuarioPuedeModificarCurso(authHeader, curso.getInstructorId())) {
+            throw new RuntimeException("No tienes permiso para modificar este módulo");
+        }
+
+       if (titulo != null && !titulo.trim().isEmpty()) {
             modulo.setTitulo(titulo);
         }
 
@@ -74,10 +91,17 @@ public class ModuloService {
         return moduloRepository.save(modulo);
     }
 
+
     // Eliminar módulo
-    public String eliminarModulo(Long id) {
+    public String eliminarModulo(String authHeader, Long id) {
         Modulo modulo = moduloRepository.findById(id)
-        .orElseThrow(() -> new RuntimeException("Módulo no encontrado Id: " + id));
+            .orElseThrow(() -> new RuntimeException("Módulo no encontrado Id: " + id));
+
+        Long instructorId = modulo.getCurso().getInstructorId();
+
+        if (!authClient.usuarioPuedeModificarCurso(authHeader, instructorId)) {
+            throw new RuntimeException("No tienes permiso para eliminar este módulo");
+        }
 
         moduloRepository.delete(modulo);
         return "Módulo eliminado";
